% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_samplesheet.R
\name{chip_samplesheet}
\alias{chip_samplesheet}
\title{Create a ChIP-seq/CUT&RUN data samplesheet for bagPipes}
\usage{
chip_samplesheet(
  Project,
  Cell_Type,
  nSamples,
  nReps,
  Genotype,
  Bio_Rep,
  Tech_Rep,
  Read1,
  Read2,
  Sequencing_Directory,
  singleEnd = FALSE
)
}
\arguments{
\item{Project}{A character value specifying which project the data is for}

\item{Cell_Type}{A character value specifying which cell type you are using}

\item{nSamples}{A numeric specifying the number of samples}

\item{nReps}{A numeric specifying the number of biological replicates}

\item{Genotype}{A character, list, or vector value specifying which cell type you are using}

\item{Bio_Rep}{A numeric specifying how many biological replicates per unique genotype}

\item{Tech_Rep}{A numeric specifying how many technical replicates per biological replicate}

\item{Read1}{A character value specifying the file path with the first .fq.gz file extension}

\item{Read2}{A character value specifying the file path with the second .fq.gz file extension}

\item{Sequencing_Directory}{A character value specifying the file path housing the .fq.gz files}

\item{singleEnd}{a logical value specifying whether or not the data is single-end sequenced}
}
\value{
data.frame, .txt file
}
\description{
Create a ChIP-seq/CUT&RUN data samplesheet for bagPipes
}
