% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietJuicerTools-package.R
\docType{package}
\name{dietJuicerTools-package}
\alias{dietJuicerTools}
\alias{dietJuicerTools-package}
\title{dietJuicerTools: This package is a suite of tools for users of the Phanstiel Lab's genomic data processing pipelines}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jpflores-13.github.io/dietJuicerTools/}
}

}
\author{
\strong{Maintainer}: JP Flores \email{jpflores013@gmail.com} (\href{https://orcid.org/0000-0001-5619-8990}{ORCID})

}
\keyword{internal}
